+++ Engine/Vec2D.h
@@ -26,7 +26,7 @@ public:
 	float GetLenght(Vec2_<T>& point1)const  {
 		const T x0 = x - point1.x;
 		const T y0 = y - point1.y;
-		return sqrt(x0 * x0 + y0 * y0);
+		return float(sqrt(x0 * x0 + y0 * y0));
 	}
 	Vec2_& operator+=(T val)
 	{
